refactor: Update Screen component props and styles
Update the Screen component to use the ComponentProps type from React and spread the rest of the props onto the wrapping div. This allows for more flexibility when using the Screen component and ensures that any additional props are correctly passed down to the underlying div element. Also, refactor the className prop to be optional and provide a default empty string value. This improves the readability and maintainability of the code.